requirements.txt
click
easyavro>=2.5.0
msgpack-python
sqlalchemy
psycopg2
